RYSK-50: Extract transaction functions to higher scope.
