+++ packages/front-end/src/components/optionsTrading/Modals/OperatorModal/index.tsx
@@ -6,11 +6,11 @@ import { Button } from "src/components/shared/Button";
 import { useGlobalContext } from "src/state/GlobalContext";
 import { ActionType } from "src/state/types";
 
+import { setOperator } from "src/components/shared/utils/transactions/setOperator";
+import { useNotifications } from "../../hooks/useNotifications";
 import { Header } from "../Shared/components/Header";
 import { Modal } from "../Shared/components/Modal";
-import { useNotifications } from "../../hooks/useNotifications";
 import { getButtonProps } from "../Shared/utils/getButtonProps";
-import { setOperator } from "../Shared/utils/transactions";
 
 export const OperatorModal = () => {
   const {
