+++ packages/front-end/src/components/optionsTrading/Modals/Shared/types.ts
@@ -8,27 +8,7 @@ export interface Addresses {
   vaultID?: string;
 }
 
-export interface AddressesRequired extends Addresses {
-  token: HexString;
-  user: HexString;
-}
-
 export interface AllowanceState {
   approved: boolean;
   amount: BigNumber;
-}
-
-export interface AddressesRequiredVaultSell extends Addresses {
-  token: HexString;
-  user: HexString;
-  collateral: HexString;
-}
-
-export interface VaultQueryData {
-  vault: {
-    id: string;
-    collateralAmount: string;
-    shortAmount: string;
-    collateralAsset: { id: string };
-  };
 }
