+++ packages/front-end/src/components/optionsTrading/Modals/Shared/utils/getButtonProps.tsx
@@ -11,7 +11,7 @@ import { capitalise } from "src/utils/caseConvert";
  * @param handleTransaction - Function to handle the transaction - optional.
  */
 export const getButtonProps = (
-  action: "buy" | "sell" | "update" | "close",
+  action: "buy" | "close" | "deposit" | "sell" | "withdraw",
   transactionPending: boolean,
   isApproved: boolean,
   handleApprove: () => Promise<void>,
