+++ packages/front-end/src/components/shared/utils/transactions/closeLong.ts
@@ -1,0 +1,62 @@
+import { BigNumber as BigNumberType } from "ethers";
+
+import { prepareWriteContract, writeContract } from "@wagmi/core";
+import { BigNumber } from "ethers";
+
+import { OptionExchangeABI } from "src/abis/OptionExchange_ABI";
+import { waitForTransactionOrTimer } from "src/components/shared/utils/waitForTransaction";
+import {
+  EMPTY_SERIES,
+  GAS_MULTIPLIER,
+  ZERO_ADDRESS,
+} from "src/config/constants";
+import OperationType from "src/enums/OperationType";
+import RyskActionType from "src/enums/RyskActionType";
+
+export const closeLong = async (
+  acceptablePremium: BigNumberType,
+  amount: BigNumberType,
+  exchangeAddress: HexString,
+  refresh: () => void,
+  tokenAddress: HexString,
+  userAddress: HexString
+) => {
+  const txData = [
+    {
+      operation: OperationType.RyskAction,
+      operationQueue: [
+        {
+          actionType: BigNumber.from(RyskActionType.SellOption),
+          owner: ZERO_ADDRESS,
+          secondAddress: userAddress,
+          asset: tokenAddress,
+          vaultId: BigNumber.from(0),
+          amount,
+          optionSeries: EMPTY_SERIES,
+          indexOrAcceptablePremium: acceptablePremium,
+          data: "0x" as HexString,
+        },
+      ],
+    },
+  ];
+
+  const config = await prepareWriteContract({
+    address: exchangeAddress,
+    abi: OptionExchangeABI,
+    functionName: "operate",
+    args: [txData],
+  });
+  config.request.gasLimit = config.request.gasLimit
+    .mul(GAS_MULTIPLIER * 100)
+    .div(100);
+
+  if (config.request.data) {
+    const { hash } = await writeContract(config);
+
+    await waitForTransactionOrTimer(hash);
+
+    refresh();
+
+    return hash;
+  }
+};
